CommitToGenerationTransactionDiffTest
